+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to eww will be listed here, starting at changes since versio
 ### BREAKING CHANGES
 - [#1176](https://github.com/elkowar/eww/pull/1176) changed safe access (`?.`) behavior:
   Attempting to index in an empty JSON string (`'""'`) is now an error.
+- `:exclusive` window parameter is now an enum (`none`, `exclusive` or `ignore`) instead of a boolean
 
 ### Fixes
 - Fix crash on invalid `formattime` format string (By: luca3s)
@@ -23,6 +24,7 @@ All notable changes to eww will be listed here, starting at changes since versio
 - Improve multi-monitor handling under wayland (By: bkueng)
 
 ### Features
+- Add `exclusive`, `ignore` and `none` to window `:exclusive` argument.
 - Add warning and docs for incompatible `:anchor` and `:exclusive` options
 - Add `eww poll` subcommand to force-poll a variable (By: kiana-S)
 - Add OnDemand support for focusable on wayland (By: GallowsDove)